feat(Python Console): add clear option to python console
ref: https://github.com/o3de/o3de/issues/11478
