+++ Code/Editor/Controls/ConsoleSCB.cpp
@@ -338,7 +338,7 @@ CConsoleSCB::CConsoleSCB(QWidget* parent)
 
     connect(ui->button, &QPushButton::clicked, this, &CConsoleSCB::showVariableEditor);
     connect(ui->findButton, &QPushButton::clicked, this, &CConsoleSCB::toggleConsoleSearch);
-    connect(ui->textEdit, &ConsoleTextEdit::searchBarRequested, this, [this]
+    connect(ui->textEdit, &AzToolsFramework::ConsoleTextEdit::searchBarRequested, this, [this]
     {
         this->ui->findBar->setVisible(true);
         this->ui->lineEditFind->setFocus();
@@ -657,96 +657,6 @@ static void OnConsoleVariableUpdated(IVariable* pVar)
     }
 }
 
-ConsoleTextEdit::ConsoleTextEdit(QWidget* parent)
-    : QPlainTextEdit(parent)
-    , m_contextMenu(new QMenu(this))
-{
-    setContextMenuPolicy(Qt::CustomContextMenu);
-    connect(this, &QPlainTextEdit::customContextMenuRequested, this, &ConsoleTextEdit::showContextMenu);
-
-    // Make sure to add the actions to this widget, so that the ShortCutDispatcher picks them up properly
-
-    QAction* copyAction = m_contextMenu->addAction(tr("&Copy"));
-    copyAction->setShortcutContext(Qt::WidgetWithChildrenShortcut);
-    copyAction->setShortcut(QKeySequence::Copy);
-    copyAction->setEnabled(false);
-    connect(copyAction, &QAction::triggered, this, &QPlainTextEdit::copy);
-    addAction(copyAction);
-
-    QAction* selectAllAction = m_contextMenu->addAction(tr("Select &All"));
-    selectAllAction->setShortcutContext(Qt::WidgetWithChildrenShortcut);
-    selectAllAction->setShortcut(QKeySequence::SelectAll);
-    selectAllAction->setEnabled(false);
-    connect(selectAllAction, &QAction::triggered, this, &QPlainTextEdit::selectAll);
-    addAction(selectAllAction);
-
-    m_contextMenu->addSeparator();
-
-    QAction* deleteAction = m_contextMenu->addAction(tr("Delete"));
-    deleteAction->setShortcutContext(Qt::WidgetWithChildrenShortcut);
-    deleteAction->setShortcut(QKeySequence::Delete);
-    deleteAction->setEnabled(false);
-    connect(deleteAction, &QAction::triggered, this, [=]() { textCursor().removeSelectedText(); } );
-    addAction(deleteAction);
-
-    QAction* clearAction = m_contextMenu->addAction(tr("Clear"));
-    clearAction->setShortcutContext(Qt::WidgetWithChildrenShortcut);
-    clearAction->setShortcut(Qt::CTRL + Qt::SHIFT + Qt::Key_C);
-    clearAction->setEnabled(false);
-    connect(clearAction, &QAction::triggered, this, &QPlainTextEdit::clear);
-    addAction(clearAction);
-
-    QAction* findAction = m_contextMenu->addAction(tr("Find"));
-    findAction->setShortcutContext(Qt::WidgetWithChildrenShortcut);
-    findAction->setShortcut(QKeySequence::Find);
-    findAction->setEnabled(true);
-    connect(findAction, &QAction::triggered, this, &ConsoleTextEdit::searchBarRequested);
-    addAction(findAction);
-
-    connect(this, &QPlainTextEdit::copyAvailable, copyAction, &QAction::setEnabled);
-    connect(this, &QPlainTextEdit::copyAvailable, deleteAction, &QAction::setEnabled);
-    connect(this, &QPlainTextEdit::textChanged, selectAllAction, [=]
-        {
-            if (document() && !document()->isEmpty())
-            {
-                clearAction->setEnabled(true);
-                selectAllAction->setEnabled(true);
-            }
-            else
-            {
-                clearAction->setEnabled(false);
-                selectAllAction->setEnabled(false);
-            }
-        });
-}
-
-bool ConsoleTextEdit::event(QEvent* theEvent)
-{
-    if (theEvent->type() == QEvent::ShortcutOverride)
-    {
-        // ignore several possible key combinations to prevent them bubbling up to the main editor
-        QKeyEvent* shortcutEvent = static_cast<QKeyEvent*>(theEvent);
-
-        QKeySequence::StandardKey ignoredKeys[] = { QKeySequence::Backspace };
-
-        for (auto& currKey : ignoredKeys)
-        {
-            if (shortcutEvent->matches(currKey))
-            {
-                // these shortcuts are ignored. Accept them and do nothing.
-                theEvent->accept();
-                return true;
-            }
-        }
-    }
-
-    return QPlainTextEdit::event(theEvent);
-}
-
-void ConsoleTextEdit::showContextMenu(const QPoint &pt)
-{
-    m_contextMenu->exec(mapToGlobal(pt));
-}
 
 ConsoleVariableItemDelegate::ConsoleVariableItemDelegate(QObject* parent)
     : QStyledItemDelegate(parent)
